requirements.txt
# Xueqiu Investor Discussion Crawler
# Python 3.8+ required

# Async HTTP and Web Scraping
aiohttp>=3.9.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Async Support
asyncio-throttle>=1.0.2
aiofiles>=23.2.0

# Configuration & Environment
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Logging & CLI
rich>=13.7.0
typer>=0.9.0

# Date & Time
python-dateutil>=2.8.2

# Hashing & Validation
xxhash>=3.4.0

# HTTP Retry
tenacity>=8.2.0

# JSON Processing
orjson>=3.9.0

# Progress Bar
tqdm>=4.66.0

# LLM API Client (Fireworks)
openai>=1.6.0
httpx>=0.25.0
